Added config for Ender 3 V2
Pasted the two configuration files from the branch release-2.1.1 into the
Configurations folder in the Marlin repo, overwriting the default config. There
are no configuration files for the 4.2.7 board but I think all it takes is
changing some strings referring to that. I hope bugfix-2.1.x is the correct
branch.